@@ -1,6 +1,6 @@
 module github.com/janghanul090801/go-backend-clean-architecture-fiber
 
-go 1.25.5
+go 1.26.0
 
 require (
 	entgo.io/ent v0.14.5
